Separate commands into invidual modules and reference courses by role
